backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
beautifulsoup4==4.12.3
lxml==5.2.2
feedparser==6.0.11
python-dotenv==1.0.1
pydantic==2.7.4
aiohttp==3.9.5
yfinance==0.2.43
apscheduler==3.10.4
textblob==0.18.0
websockets==12.0
python-multipart==0.0.9
aiosqlite==0.20.0
pytz==2024.1
requests==2.32.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
